Refactor transform.js into a ScriptTransformer
