+++ packages/jest-cli/src/generateEmptyCoverage.js
@@ -14,7 +14,7 @@ import type {ProjectConfig, Path} from 'types/Config';
 
 const IstanbulInstrument = require('istanbul-lib-instrument');
 
-const {transformSource, shouldInstrument} = require('jest-runtime');
+const {ScriptTransformer, shouldInstrument} = require('jest-runtime');
 
 module.exports = function(
   source: string,
@@ -24,7 +24,11 @@ module.exports = function(
   if (shouldInstrument(filename, config)) {
     // Transform file without instrumentation first, to make sure produced
     // source code is ES6 (no flowtypes etc.) and can be instrumented
-    const transformResult = transformSource(filename, config, source, false);
+    const transformResult = new ScriptTransformer(config).transformSource(
+      filename,
+      source,
+      false,
+    );
     const instrumenter = IstanbulInstrument.createInstrumenter();
     instrumenter.instrumentSync(transformResult.code, filename);
     return {
