+++ packages/jest-runtime/src/__tests__/instrumentation-test.js
@@ -21,14 +21,16 @@ const FILE_PATH_TO_INSTRUMENT = path.resolve(
 
 it('instruments files', () => {
   const vm = require('vm');
-  const transform = require('../transform');
+  const ScriptTransformer = require('../ScriptTransformer');
   const config = {
     cache: false,
     cacheDirectory: os.tmpdir(),
     collectCoverage: true,
     rootDir: '/',
   };
-  const instrumented = transform(FILE_PATH_TO_INSTRUMENT, config).script;
+  const instrumented = new ScriptTransformer(config).transform(
+    FILE_PATH_TO_INSTRUMENT,
+  ).script;
   expect(instrumented instanceof vm.Script).toBe(true);
   // We can't really snapshot the resulting coverage, because it depends on
   // absolute path of the file, which will be different on different
